Core.Render

Description

描画に用いるRenderの定義などです。 ついでにCore.Viewの位置補正なども行います。

Render

data Render a

レンダリングするMonadです。Affineなのでアフィン変換もできます。

fill ∷ Color → Shape () → Render ()

fill c sで、色cで図形sを塗りつぶします。

stroke ∷ E → Color → Shape () → Render ()

stroke w c sで、色cで図形sの輪郭を幅wで描画します。

clip ∷ Shape () → Render () → Render ()

clip s aで、図形sの領域内のみに描画するようにした状態でaを実行します。

shadowed ∷ E2 → E → Color → Render () → Render ()

shadowed p d c actで、位置pに深さdの色cの影がついた状態でactを実行します。

withValue

Arguments

∷ Value a
取り出す対象のValue

→ (a → Render b)
取り出した値を使って行う処理

→ Render (Maybe b)
取り出しに失敗するとNothingが帰ります。

Value'から値を取り出して処理を行います。

Color

type Color = (E, E, E, E)

色の指定に使われる型です。R^4でEasingしてv4でColorを作る、などすると良いとおもいます。

black ∷ Color

white ∷ Color

Canvas

getCanvasById ∷ MonadIO m ⇒ String → m (Maybe Canvas)

Create a 2D drawing context from a DOM element identified by its ID.

data Canvas

A canvas; a viewport into which a picture can be rendered. The origin of the coordinate system used by the canvas is the top left corner of the canvas element. JS representation is a reference to the backing canvas element.

レンダリング

render ∷ Canvas → Render () → IO ()

描画します。

画像

data Image

画像を表現します。

getImage ∷ String → IO Image

画像をURLから取得します。何度呼び出してもコスト掛からないので好きなときに使えば良いと思います。

sizeOf ∷ MonadIO m ⇒ Image → m (Double, Double)

画像のサイズを取得します。

draw ∷ SimpleGetter Image (Render ())

画像を左上の点を原点に描画します。Lensを使ってください。

image <- getImage "res/image.png"
render $ do
  image^.draw

drawClipped ∷ E2 → E2 → SimpleGetter Image (Render ())

drawClipped p sで、左上の点をpでサイズをsとした矩形領域の画像を、左上の点を原点に描画するLensになります。

atlas <- getImage "res/atlas.png"
w <- po . (/4) . fst <$> sizeOf atlas
h <- po . snd <$> sizeOf atlas
render $ forM_ [0..3] $ i -> do
  translate (i*w) 0 $ atlas^.drawClipped (i*w,0) (w,h)

キャッシュ

cache

Arguments

∷ HashableIO a

⇒ String
キャッシュに付ける一意な名前

→ E2
領域の左上座標

→ E2
領域のサイズ

→ E2
キャッシュの解像度

→ a
制御値

→ Render ()
描画内容

→ Render ()

描画結果のキャッシュを行います。制御値に変化があると再描画します。

View関連

applyTransform ∷ Render () → World ()

Viewをレンダリングしたときの座標を取得し、実際の位置を移動させます。 使うときは、renderに渡したものと同じRender ()を毎フレームapplyTransformにも渡せば良いです。