Core.Util

Description

とりあえずimportしておけばいいモジュールです。 よく使う便利な関数とかが入ってます。 ほしいものがあったら好きに追加してください。

Exports

Lens, State Monadなどの良く使うやつをexportしています。

Haste.Concurrent

class Monad m ⇒ MonadConc m where

Any monad which supports concurrency.

Minimal complete definition

liftConc, fork

Methods

liftConc ∷ CIO a → m a

fork ∷ m () → m ()

concurrent ∷ CIO () → IO ()

Run a concurrent computation. Two different concurrent computations may share MVars; if this is the case, then a call to concurrent may return before all the threads it spawned finish executing.

data MVar a

newMVar ∷ MonadIO m ⇒ a → m (MVar a)

Create a new MVar with an initial value.

newEmptyMVar ∷ MonadIO m ⇒ m (MVar a)

Create a new empty MVar.

takeMVar ∷ MVar a → CIO a

Read an MVar. Blocks if the MVar is empty. Only the first writer in the write queue, if any, is woken.

putMVar ∷ MVar a → a → CIO ()

Write an MVar. Blocks if the MVar is already full. Only the first reader in the read queue, if any, is woken.

peekMVar ∷ MonadIO m ⇒ MVar a → m (Maybe a)

Peek at the value inside a given MVar, if any, without removing it.

readMVar ∷ MVar a → CIO a

Read an MVar then put it back. As Javascript is single threaded, this function is atomic. If this ever changes, this function will only be atomic as long as no other thread attempts to write to the MVar.

type Outbox = MBox Send

type Inbox = MBox Recv

spawn ∷ MonadConc m ⇒ (Inbox a → m ()) → m (Outbox a)

Creates a generic process and returns a MBox which may be used to pass messages to it. While it is possible for a process created using spawn to transmit its inbox to someone else, this is a very bad idea; don't do it.

(!) ∷ MonadConc m ⇒ Outbox a → a → m ()

Write a value to a MBox. Named after the Erlang message sending operator, as both are intended for passing messages to processes. This operation does not block until the message is delivered, but returns immediately.

(<!) ∷ MonadConc m ⇒ Outbox a → m a → m ()

Perform a Client computation, then write its return value to the given pipe. Mnemonic: the operator is a combination of <- and !. Just like (!), this operation is non-blocking.

receive ∷ MonadConc m ⇒ Inbox a → m a

Block until a message arrives in a mailbox, then return it.

よく使う型

type Z = Integer

type R = Double

type R2 = (R, R)

IO系

io ∷ MonadIO m ⇒ IO a → m a

IO Actionを持ち上げます。 MonadIO制約は使うときは大体はm = Worldだとおもいます。

>>> io $ print 42
42

makeMVar ∷ MonadIO m ⇒ a → m (MVar a)

簡単に言うと変数を作ります。

withMVar ∷ (MonadIO m, MonadConc m) ⇒ MVar a → StateT a m b → m b

変数の中身を取り出した状態で計算を行います。 Stata Monad上で動くのでLensでうまいこと操作したりlift関数で持ち上げたりすることになります。

do
  x <- makeMVar 0
  withMVar x $ do
    id += 5
  x' <- liftW $ readMVar x
  io $ print x' -- 5

Monad系

(??) ∷ Functor f ⇒ f (a → b) → a → f b

runStateTするときに便利だけど使わないかも

onJust ∷ Monad m ⇒ Maybe a → (a → m ()) → m Bool

Justのときに処理を行うときに便利。

ループ系

while ∷ Monad m ⇒ m Bool → m ()

条件を満たす間ループします。

loop

Arguments

∷ MonadCont m

⇒ a
初期値

→ ((b → m c) → a → m a)
exit関数と前回の状態を受取り、次の状態を返す関数

→ m b
結果 : exitに渡された引数

引数に渡されたexit関数を呼ぶまで処理をループします。m = Worldだとおもいます、だいたい。

waitFor ∷ Monad m ⇒ (a → Bool) → m a → m a

条件を満たす値が来るまでループします。満たした値を返すのがwhileとの違い。

条件式

基本的には条件ラムダ式をそのまま論理演算するためのものです。

(==3) |?| (==5) ≡ \x -> x == 3 || x == 5

(&?&) ∷ Applicative m ⇒ m Bool → m Bool → m Bool infixl 3

(|?|) ∷ Applicative m ⇒ m Bool → m Bool → m Bool infixl 3

添字付きループ

添字付きでforMを使いたいときのための関数です。

>>> iforM_ [2,5,6] $ \i x -> print (i,x)
(0,2)
(1,5)
(2,6)

iforM ∷ Monad m ⇒ [a] → (Int → a → m b) → m [b]

iforM_ ∷ Monad m ⇒ [a] → (Int → a → m b) → m ()