Haste.Events

Description

Event handling for Haste.

Core event functionality

class Event evt where

Any type that describes an event.

Minimal complete definition

eventName, eventData

Associated Types

type EventData evt

The type of data to pass to handlers for this event.

Methods

eventName ∷ evt → JSString

The name of this event, as expected by the DOM.

eventData ∷ evt → JSAny → IO (EventData evt)

Construct event data from the event identifier and the JS event object.

class MonadIO m ⇒ MonadEvent m where

Any monad in which we're able to handle events.

Minimal complete definition

mkHandler

Methods

mkHandler ∷ (a → m ()) → m (a → IO ())

data HandlerInfo

Information about an event handler.

unregisterHandler ∷ MonadIO m ⇒ HandlerInfo → m ()

Unregister an event handler.

onEvent

Arguments

∷ (MonadEvent m, IsElem el, Event evt)

⇒ el
Element to set handler on.

→ evt
Event to handle.

→ (EventData evt → m ())
Event handler.

→ m HandlerInfo
Information about the handler.

Set an event handler on a DOM element.

preventDefault ∷ IO ()

Prevent the event being handled from resolving normally. Does nothing if called outside an event handler.

Basic events with no arguments

data BasicEvent

Constructors

Load

Unload

Change

Focus

Blur

Submit

Scroll

Keyboard-related events

data KeyEvent

Constructors

KeyPress

KeyUp

KeyDown

data KeyData

Event data for keyboard events.

Constructors

KeyData

Fields

keyCode ∷ !Int

keyCtrl ∷ !Bool

keyAlt ∷ !Bool

keyShift ∷ !Bool

keyMeta ∷ !Bool

mkKeyData ∷ Int → KeyData

Build a KeyData object for the given key, without any modifier keys pressed.

Mouse-related events

data MouseEvent

Constructors

Click

DblClick

MouseDown

MouseUp

MouseMove

MouseOver

MouseOut

Wheel

data MouseData

Event data for mouse events.

Constructors

MouseData

Fields

mouseCoords ∷ !(Int, Int)

Mouse coordinates.

mouseButton ∷ !(Maybe MouseButton)

Pressed mouse button, if any.

mouseWheelDeltas ∷ !(Double, Double, Double)

(x, y, z) mouse wheel delta. Always all zeroes except for Wheel.

data MouseButton

Constructors

MouseLeft

MouseMiddle

MouseRight

Touch-related events

data TouchEvent

Constructors

TouchStart

TouchMove

TouchEnd

TouchCancel

data TouchData

Event data for touch events.

Constructors

TouchData

Fields

touches ∷ [Touch]

All fingers currently on the screen.

targetTouches ∷ [Touch]

All fingers on the DOM element receiving the event.

changedTouches ∷ [Touch]

All fingers involved in the current event. For instance, fingers removed for the TouchEnd event.

data Touch

A finger touching the screen.

Constructors

Touch

Fields

identifier ∷ !Int

Unique identifier for this touch.

target ∷ !Elem

Target element of the touch.

pageCoords ∷ !(Int, Int)

Page coordinates of the touch, including scroll offsets.

clientCoords ∷ !(Int, Int)

Page coordinates of the touch, excluding scroll offsets.

screenCoords ∷ !(Int, Int)

Screen coordinates of the touch.