Haste.Hash

Description

Hash manipulation and callbacks.

Documentation

onHashChange ∷ MonadIO m ⇒ (String → String → IO ()) → m ()

Register a callback to be run whenever the URL hash changes. The first and second argument of the callback are the old and new and hash respectively.

onHashChange' ∷ MonadIO m ⇒ (JSString → JSString → IO ()) → m ()

JSString version of onHashChange.

setHash ∷ MonadIO m ⇒ String → m ()

Set the hash part of the current URL.

getHash ∷ MonadIO m ⇒ m String

Read the hash part of the currunt URL.

setHash' ∷ MonadIO m ⇒ JSString → m ()

Set the hash part of the current URL - JSString version.

getHash' ∷ MonadIO m ⇒ m JSString

Read the hash part of the currunt URL - JSString version.